fix: improvements
