+++ src/components/Header/hooks.ts
@@ -1,6 +1,6 @@
+import { COLLECTION_PATHNAME } from '@/constants/navigation'
 import { useCurrentProductMedia, useGetAllProductThemeMedia } from '@/state/collection/hooks'
 import { ShopImageSrcSet } from '@/types'
-import { checkIsCollectionPage } from '@/utils/navigation'
 import { BLACK } from '@past3lle/theme'
 import { getIsMobile } from '@past3lle/utils'
 import { usePathname } from 'next/navigation'
@@ -11,28 +11,34 @@ interface ReturnType {
   color: string
 }
 export function useHeaderMedia(): ReturnType {
-  const pathname = usePathname() || ''
-  const isCollectionPage = checkIsCollectionPage({ pathname })
-  const dynamicMediaConfig = useCurrentProductMedia()
+  const pathname = usePathname() || '/'
+  const currentOnScreenMetaAsset = useCurrentProductMedia(pathname)
   // "randomly" select a product header from collection for header (on mobile collection view ONLY)
-  const productMediaList = useGetAllProductThemeMedia()
+  const productsMetaAssetsList = useGetAllProductThemeMedia()
   const [{ headerLogo, color }, setHeaderAssets] = useState<ReturnType>({ headerLogo: undefined, color: BLACK })
   // Wait for client to be ready to set assets on header as it is SSR
   useEffect(() => {
-    if (productMediaList) {
-      const randomIndex = Math.floor(Math.random() * productMediaList.length - 1)
-      const headerLogo =
-        getIsMobile() && isCollectionPage
-          ? productMediaList?.[randomIndex]?.headerLogo
-          : dynamicMediaConfig?.headerLogoSet
-      setHeaderAssets({
-        headerLogo,
-        color: dynamicMediaConfig?.color ?? BLACK,
-      })
-    }
+    const isCollectionPage = pathname === COLLECTION_PATHNAME
+    const isHomePage = pathname === '/'
+
+    const randomIndex = Math.floor(Math.random() * (productsMetaAssetsList?.length || 0) - 1)
+    const headerLogo =
+      getIsMobile() && isCollectionPage
+        ? productsMetaAssetsList?.[randomIndex]?.headerLogo
+        : currentOnScreenMetaAsset?.headerLogoSet
+    setHeaderAssets({
+      headerLogo,
+      color: (isHomePage ? currentOnScreenMetaAsset?.color : currentOnScreenMetaAsset?.color) ?? BLACK,
+    })
+
     // Can safely ignore here as we only want the media list length
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [dynamicMediaConfig?.color, dynamicMediaConfig?.headerLogoSet, isCollectionPage, productMediaList?.length])
+  }, [
+    currentOnScreenMetaAsset?.color,
+    currentOnScreenMetaAsset?.headerLogoSet,
+    pathname,
+    productsMetaAssetsList?.length,
+  ])
 
   return {
     headerLogo,
