+++ src/pages/index.tsx
@@ -1,50 +1,16 @@
 /* eslint-disable react-refresh/only-export-components */
-import { SkillLockStatus } from '@past3lle/forge-web3'
-import { useIsMobile } from '@past3lle/hooks'
-import { OFF_WHITE } from '@past3lle/theme'
 import SEO from '@/components/SEO'
 import { SelectedShowcaseVideo } from '@/components/Showcase/Videos'
-import { COLLECTION_PATHNAME } from '@/constants/navigation'
-import {
-  SingleProductAsidePanel,
-  SingleProductContainer,
-  SingleProductScreensContainer,
-} from '@/components/Asides/skill/styled'
-import { CTA_BUTTON_PROP_THEMES } from '@/components/PagesComponents/constants'
-import { useGetCommonPropsFromProduct } from '@/components/PagesComponents/hooks/useGetCommonPropsFromProduct'
-import { useProductWebCarouselActions } from '@/components/PagesComponents/hooks/useProductCarouselActions'
-import { ProductSubHeader } from '@/components/PagesComponents/styleds'
-import { useCallback, useMemo } from 'react'
-import { getImageSizeMap, mapShopifyHomepageToProps } from '@/shopify/utils'
-import { useThemeManager } from '@/state/user/hooks'
-import { BLACK_TRANSPARENT_MORE } from '@/theme'
+import { SingleProductContainer } from '@/components/Asides/skill/styled'
+import { mapShopifyHomepageToProps } from '@/shopify/utils'
 import dynamic from 'next/dynamic'
-import { useRouter } from 'next/router'
 import { homeQuery } from '@/shopify/graphql/api/home'
 import { ProductSchema } from '@/components/SEO/types'
 import { getProductSeoSchema } from '@/components/SEO/utils'
 import { AnimatedPastelleLoader } from '@/components/Loader'
-
-const AsideCarousel = dynamic(
-  () =>
-    import(
-      /* webpackPrefetch: true,  webpackChunkName: "ASIDECAROUSEL" */ '@/components/PagesComponents/screens/AsideCarousel'
-    )
-)
-const ActionScreen = dynamic(
-  () =>
-    import(
-      /* webpackPrefetch: true,  webpackChunkName: "ACTIONSCREEN" */ '@/components/PagesComponents/screens/ActionsAndChildren'
-    ),
-  { ssr: false }
-)
-const Description = dynamic(
-  () =>
-    import(
-      /* webpackPrefetch: true,  webpackChunkName: "DESCRIPTION" */ '@/components/PagesComponents/screens/DescriptionAndChildren'
-    ),
-  { ssr: false }
-)
+import HomepageMainContent from '@/components/Homepage/HomePageClientContent'
+import { wrapper } from '@/state'
+import { updateCurrentlyViewing, updateHomepage } from '@/state/collection/reducer'
 
 const SinglePageSmartWrapper = dynamic(
   () => import(/* webpackPrefetch: true,  webpackChunkName: "HOMESMARTWRAPPER" */ '@/components/PagesComponents'),
@@ -65,57 +31,29 @@ const VIDEO_CSS = `
   }
 `
 
-const OVERSIZED_CAROUSEL_MEDIA_WIDTH = '140%'
-
 // This also gets called at build time
-export async function getStaticProps() {
+export const getStaticProps = wrapper.getStaticProps((store) => async (): Promise<{ props: Props }> => {
   const { product: homeRaw } = await homeQuery()
 
   if (!homeRaw) throw new Error('Missing homepage data!')
   const homepage = homeRaw ? mapShopifyHomepageToProps(homeRaw) : null
 
+  if (homepage) {
+    store.dispatch(updateHomepage(homepage))
+    store.dispatch(updateCurrentlyViewing({ handle: 'pastelle', id: 'gid://shopify/Product/8818646909251' }))
+  }
+
   // Pass post data to the page via props
   return { props: { homepage, schemaSEO: getProductSeoSchema(homepage) } }
-}
+})
 
 type Props = {
-  homepage: ReturnType<typeof mapShopifyHomepageToProps>
+  homepage: ReturnType<typeof mapShopifyHomepageToProps> | null
   schemaSEO: ProductSchema | null
 }
 
 export default function Home({ homepage, schemaSEO }: Props) {
-  const { push: navigate } = useRouter()
-
-  const { mode } = useThemeManager()
-  const { onChange } = useProductWebCarouselActions({
-    startIndex: 0,
-  })
-
-  const isMobile = useIsMobile()
-  // Get common props for screens
-  // lockStatus === null means ignore skill state
-  const commonProps = useGetCommonPropsFromProduct({
-    ...homepage,
-    bgColor: homepage?.bgColor || BLACK_TRANSPARENT_MORE,
-    color: homepage?.color || OFF_WHITE,
-    altColor: homepage?.altColor || BLACK_TRANSPARENT_MORE,
-    lockedImages: [],
-    sizeChart: [],
-    title: homepage?.title || 'PASTELLE',
-    handle: homepage?.handle || 'PASTELLE',
-    productType: '',
-    variant: null,
-    isMobile,
-    lockStatus: SkillLockStatus.OWNED,
-  })
-
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const navigateToCollection = useCallback(() => navigate(COLLECTION_PATHNAME), [])
-
-  // Src-set of all images
-  const imageSrcSet = useMemo(() => getImageSizeMap(homepage?.images || []), [homepage?.images])
-
-  if (!homepage?.images?.length || !homepage?.videos?.length || !schemaSEO) return null
+  if (!homepage || !schemaSEO) return null
 
   return (
     <>
@@ -128,108 +66,13 @@ export default function Home({ homepage, schemaSEO }: Props) {
         schema={schemaSEO}
       />
       <SinglePageSmartWrapper>
-        {({
-          screensContainerNode,
-          rootContainerNode,
-          asideContainerRef,
-          setScreensContainerRef,
-          parentAspectRatio,
-        }) => (
+        {({ parentAspectRatio, ...restProps }) => (
           <SingleProductContainer id="#item-container" parentAspectRatio={parentAspectRatio}>
-            <SingleProductAsidePanel id="#item-aside-panel" ref={asideContainerRef}>
-              <SingleProductScreensContainer
-                $calculatedSizes={{
-                  height: asideContainerRef.current?.clientHeight,
-                  width: screensContainerNode?.clientWidth,
-                }}
-                ref={setScreensContainerRef}
-                bgColor={homepage?.bgColor}
-                navLogo={homepage?.navLogo}
-                logo={homepage?.logo}
-              >
-                <AsideCarousel
-                  {...commonProps}
-                  containerNode={screensContainerNode}
-                  carousel={{
-                    touchAction: 'pan-y',
-                    data: commonProps.isMobile ? [...homepage.videos, ...imageSrcSet] : imageSrcSet,
-                    imageProps: {
-                      style: {
-                        maxWidth: OVERSIZED_CAROUSEL_MEDIA_WIDTH,
-                      },
-                    },
-                    videoProps: {
-                      height: 'auto',
-                      width: OVERSIZED_CAROUSEL_MEDIA_WIDTH,
-                    },
-                    startIndex: 0,
-                    onChange,
-                    onCarouselItemClick: navigateToCollection,
-                    indicatorOptions: {
-                      showIndicators: true,
-                    },
-                  }}
-                  themeMode={mode}
-                  breadcrumbs={{
-                    breadcrumbs: ['PASTELLE APPAREL'],
-                    lastCrumb: '',
-                  }}
-                  skillInfo={null}
-                  userAddress={undefined}
-                  logoCss="filter: unset;"
-                />
-                <ActionScreen
-                  {...commonProps}
-                  labels={{
-                    main: (isMobile ? 'TAP TO ' : '') + 'VIEW COLLECTION',
-                    async: 'HEADING TO COLLECTION',
-                  }}
-                  skillInfo={null}
-                  callback={navigateToCollection}
-                  fixedWidth={screensContainerNode?.clientWidth}
-                  rootNode={rootContainerNode}
-                  fixedButtonStyles={CTA_BUTTON_PROP_THEMES.GLACIUS}
-                  staticButtonStyles={CTA_BUTTON_PROP_THEMES.GLACIUS}
-                >
-                  <ProductSubHeader
-                    fontWeight={100}
-                    fontStyle="none"
-                    css={`
-                      > div {
-                        width: max-content;
-                        background-color: ${homepage.bgColor};
-                        &:not(:last-child) {
-                          margin: 0 0 4px;
-                        }
-                        padding: 2px 8px;
-                      }
-                    `}
-                  >
-                    <div>
-                      <strong>GENES1S</strong> HEAVYWEIGHT COLLECTION
-                    </div>
-                    <div>AVAILABLE NOW</div>
-                  </ProductSubHeader>
-                </ActionScreen>
-                <Description
-                  {...commonProps}
-                  header=""
-                  containerNode={screensContainerNode}
-                  description={homepage.description}
-                />
-              </SingleProductScreensContainer>
-            </SingleProductAsidePanel>
+            <HomepageMainContent {...restProps} homepage={homepage} parentAspectRatio={parentAspectRatio} />
             <SelectedShowcaseVideo
               forceLoad
               firstPaintOver
               isMobileWidth={false}
-              smartFill={
-                isMobile
-                  ? undefined
-                  : {
-                      full: true,
-                    }
-              }
               hideVideo={false}
               css={VIDEO_CSS}
               currentCarouselIndex={0}
